Extract shared AlphaGenome API plumbing into alphagenome_api
Move the grpc retry wrappers and splice-track lookup out of
alphagenome_calibration into a new alphagenome_api module, so the calibration
code (and later the deletion experiment) share one copy. Pure refactor: the
calibration algorithm is unchanged.
